refactor: reorganize normalization functions +
- reorganize normalization functions
- split validate-and-parse import so patient conversion has
  it's own file

Ref. metriport/metriport-internal#2330
