+++ packages/core/src/command/patient-import/csv/contact.ts
@@ -1,12 +1,7 @@
-import {
-  errorToString,
-  isEmailValid,
-  isPhoneValid,
-  normalizeEmail,
-  normalizePhoneNumber,
-} from "@metriport/shared";
+import { errorToString, normalizeEmailNewSafe, normalizePhoneNumberSafe } from "@metriport/shared";
 import { filterTruthy } from "@metriport/shared/common/filter-map";
 import { Contact } from "../../../domain/contact";
+import { out } from "../../../util";
 import { ParsingError } from "./shared";
 
 const maxContacts = 10;
@@ -15,6 +10,7 @@ export function mapCsvContacts(csvPatient: Record<string, string>): {
   contacts: Contact[];
   errors: ParsingError[];
 } {
+  const { log } = out(`mapCsvContacts`);
   const errors: ParsingError[] = [];
   const contacts: (Contact | undefined)[] = [];
 
@@ -30,10 +26,8 @@ export function mapCsvContacts(csvPatient: Record<string, string>): {
 
   const filteredContacts = contacts.flatMap(filterTruthy);
   if (filteredContacts.length > maxContacts) {
-    errors.push({
-      field: "address",
-      error: `Found more than ${maxContacts} contacts`,
-    });
+    log(`Found more than ${maxContacts} contacts, discarding the rest`);
+    filteredContacts.splice(maxContacts);
   }
 
   return { contacts: filteredContacts, errors };
@@ -50,14 +44,16 @@ function parseContact(
 
   let email: string | undefined = undefined;
   try {
-    email = normalizeEmailUtils(csvPatient[emailName]);
+    email = normalizeEmail(csvPatient[emailName]);
+    if (!email) throw new Error(`Invalid ${emailName}`);
   } catch (error) {
     errors.push({ field: emailName, error: errorToString(error) });
   }
 
   let phone: string | undefined = undefined;
   try {
-    phone = normalizePhoneNumberUtils(csvPatient[phoneName]);
+    phone = normalizePhoneNumber(csvPatient[phoneName]);
+    if (!phone) throw new Error(`Invalid ${phoneName}`);
   } catch (error) {
     errors.push({ field: phoneName, error: errorToString(error) });
   }
@@ -71,18 +67,14 @@ function parseContact(
   return { contact, errors };
 }
 
-export function normalizePhoneNumberUtils(phone: string | undefined): string | undefined {
+export function normalizePhoneNumber(phone: string | undefined): string | undefined {
   if (phone == undefined) return undefined;
-  const normalPhone = normalizePhoneNumber(phone);
-  if (normalPhone.length === 0) return undefined;
-  if (!isPhoneValid(normalPhone)) throw new Error("Invalid Phone");
+  const normalPhone = normalizePhoneNumberSafe(phone);
   return normalPhone;
 }
 
-export function normalizeEmailUtils(email: string | undefined): string | undefined {
+export function normalizeEmail(email: string | undefined): string | undefined {
   if (email == undefined) return undefined;
-  const normalEmail = normalizeEmail(email);
-  if (normalEmail.length === 0) return undefined;
-  if (!isEmailValid(normalEmail)) throw new Error("Invalid Email");
+  const normalEmail = normalizeEmailNewSafe(email);
   return normalEmail;
 }
