+++ packages/shared/src/domain/address/city.ts
@@ -1,0 +1,6 @@
+import { toTitleCase } from "../../common/title-case";
+
+export function normalizeCity(city: string): string | undefined {
+  if (city == undefined) return undefined;
+  return toTitleCase(city).trim();
+}
